extracting error view to its own file
