+++ site/src/kiosk/kiosk-component.jsx
@@ -3,6 +3,7 @@ import { connect } from 'react-redux'
 import * as UseCase from './kiosk-use-case'
 import * as Actions from './kiosk-actions'
 import { ContentView } from '../common/viewer/content-view'
+import ErrorView from './error-view'
 
 import * as firebase from 'firebase/app'
 import 'firebase/auth'
@@ -69,47 +70,5 @@ const mapDispatchToProps = (dispatch) => {
         resetDeviceContent: UseCase.resetDeviceContent(dispatch)
     }
 }
-
-const container = {
-    background: '#0000aa',
-    color: '#ffffff',
-    fontFamily: 'Lucida, monospace',
-    fontSize: '30pt',
-    textAlign: 'center',
-    height: '100vh',
-    width: '100vw'
-}
-
-const code = {
-    background: '#fff',
-    color: '#0000aa',
-    padding: '2px 8px',
-    fontWeight: 'bold'
-}
-
-const content = {
-    paddingLeft: '10vw',
-    paddingRight: '10vw',
-    fontSize: '15pt',
-    wordWrap: 'break-word'
-}
-
-const ErrorView = ({ error }) => (
-    <div style={container}>
-        <div style={{ paddingTop: '10vh' }}>
-            <span style={code}>ERROR {error.code}</span>
-            <p>
-                Plugin failed to provide content.
-            </p>
-            <p style={content}>{error.cause}</p>
-            <p style={content}>
-                {error.source.split('?')[0]}
-            </p>
-            <p>
-            </p>
-            <p>#dashboards / #g-it</p>
-        </div>
-    </div>
-)
 
 export const KioskComponent = connect(mapStateToProps, mapDispatchToProps)(Component)
